add ShuffledDeckGenerator
this generator will shuffle the array you give it and then yield each element from the array until it reaches the end of the array. at that point it will reshuffle the array and reset its pointer to the beginning of the array and continue yielding forever. (like a deck of cards)