add TaskRunner and associated sanity test
no real tests yet. i'm not entirely sure how i want to break it down for testing or better illustrating the point of the TaskRunner. basically it's just a synchronous queue to stack up a bunch of commands. it's almost an unnecessary extraction from the root class, but i have plans to allow the higher-level, public classes chain together tasks more freely and this hopefully should facilitate that.